+++ PyAnalySeries.py
@@ -48,7 +48,7 @@ else:
     filesName = None
 
 #========================================================================================
-version = 'v5.13'
+version = 'v5.14'
 
 open_ws = {}
 open_displayWindows = {}