+++ releases.md
@@ -1,3 +1,6 @@
+ * 5.14
+   * Fix style for buttons
+   * Change Import to allow empty cells in 1st column when importing series
  * 5.13
    * Add last recent dir when open worksheets
    * Fix when xlsx file has no item detected